Split parser class into smaller classes
