+++ src/main/java/darkpool/command/DeleteCommand.java
@@ -1,5 +1,6 @@
 package darkpool.command;
 
+import darkpool.util.DarkpoolException;
 import darkpool.util.Storage;
 import darkpool.util.TaskList;
 import darkpool.gui.Gui;
@@ -28,7 +29,10 @@ public class DeleteCommand extends Command {
      * @return
      */
     @Override
-    public String execute(TaskList tasks, Gui gui, Storage storage) {
+    public String execute(TaskList tasks, Gui gui, Storage storage) throws DarkpoolException {
+        if (index < 0 || index >= tasks.getSize()) {
+            throw new DarkpoolException("do you know how to count? the task number is out of range");
+        }
         String delString = tasks.deleteTask(index);
         return gui.delete(delString);
     }
