+++ src/main/java/darkpool/parser/FindParser.java
@@ -1,0 +1,16 @@
+package darkpool.parser;
+
+import darkpool.command.Command;
+import darkpool.command.FindCommand;
+import darkpool.util.DarkpoolException;
+
+import java.util.Objects;
+
+public class FindParser {
+
+    static public Command parse(String[] userInput) throws DarkpoolException {
+        checkCommand.findChecker(userInput);
+        return new FindCommand(userInput[1]);
+    }
+
+}
