+++ src/main/java/darkpool/parser/ListParser.java
@@ -1,0 +1,13 @@
+package darkpool.parser;
+
+import darkpool.command.Command;
+import darkpool.command.ListCommand;
+import darkpool.util.DarkpoolException;
+
+public class ListParser {
+
+    static public Command parse() throws DarkpoolException {
+        return new ListCommand();
+    }
+
+}
