+++ src/main/java/darkpool/parser/TodoParser.java
@@ -1,0 +1,14 @@
+package darkpool.parser;
+
+import darkpool.command.AddCommand;
+import darkpool.command.Command;
+import darkpool.util.DarkpoolException;
+
+
+public class TodoParser {
+
+    static public Command parse(String[] userInput) throws DarkpoolException {
+        return new AddCommand(checkCommand.todoChecker(userInput));
+    }
+
+}
