+++ src/main/java/darkpool/parser/UnmarkParser.java
@@ -1,0 +1,18 @@
+package darkpool.parser;
+
+import darkpool.command.Command;
+import darkpool.command.UnmarkCommand;
+import darkpool.util.DarkpoolException;
+
+
+public class UnmarkParser {
+
+    static public Command parse(String[] userInput) throws DarkpoolException {
+        return new UnmarkCommand(taskAction(userInput));
+    }
+
+    private static int taskAction(String[] userInput) throws DarkpoolException {
+        return checkCommand.markUnmarkDeleteChecker(userInput);
+    }
+
+}
